+++ src/docs/pages/textarea-demo/textarea-demo.ts
@@ -9,4 +9,8 @@ export class TextareaDemo {
     required: boolean;
     disabled: boolean;
     message: string;
+    validatedMessage: string;
+
+    pattern: string = '^[0-9]*\n*[0-9]*$';
+    validationErrorTooltip: string = 'Please use the correct format';
 }